chore: upgrade Next.js to version 16.0.1 across all packages; refactor proxy middleware and update API route configurations
