+++ apps/web/package.json
@@ -108,8 +108,8 @@
     "@eslint/js": "^9.37.0",
     "@ncthub/eslint-config": "workspace:*",
     "@ncthub/typescript-config": "workspace:*",
-    "@next/bundle-analyzer": "15.3.3",
-    "@next/eslint-plugin-next": "15.3.3",
+    "@next/bundle-analyzer": "16.0.1",
+    "@next/eslint-plugin-next": "16.0.1",
     "@tanstack/eslint-plugin-query": "^5.91.0",
     "@tanstack/react-query-devtools": "^5.90.2",
     "@testing-library/react": "^16.3.0",
