+++ apps/web/src/app/api/ai/chat/google-vertex/route.ts
@@ -1,9 +1,7 @@
 import { POST } from '@ncthub/ai/chat/google-vertex/route';
 
-export const config = {
-  maxDuration: 90,
-  preferredRegion: 'sin1',
-  runtime: 'edge',
-};
+export const runtime = 'edge';
+export const maxDuration = 60;
+export const preferredRegion = 'sin1';
 
 export { POST };
