+++ apps/web/src/app/api/ai/chat/openai/summary/route.ts
@@ -1,9 +1,7 @@
 import { PATCH } from '@ncthub/ai/chat/openai/summary/route';
 
-export const config = {
-  maxDuration: 90,
-  preferredRegion: 'sin1',
-  runtime: 'edge',
-};
+export const runtime = 'edge';
+export const maxDuration = 60;
+export const preferredRegion = 'sin1';
 
 export { PATCH };
