+++ packages/ai/package.json
@@ -44,7 +44,7 @@
     "@ncthub/utils": "workspace:*",
     "ai": "^4.3.19",
     "eslint": "^9.37.0",
-    "next": "15.3.3",
+    "next": "16.0.1",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
     "zod": "^3.25.76"
