+++ packages/auth/package.json
@@ -19,7 +19,7 @@
     "@ncthub/types": "workspace:*",
     "@ncthub/ui": "workspace:*",
     "@ncthub/utils": "workspace:*",
-    "next": "15.3.3",
+    "next": "16.0.1",
     "react": "^19.2.0"
   },
   "devDependencies": {
