+++ packages/supabase/package.json
@@ -16,7 +16,7 @@
     "@supabase/supabase-js": "^2.75.0",
     "@ncthub/types": "workspace:*",
     "eslint": "^9.37.0",
-    "next": "15.3.3",
+    "next": "16.0.1",
     "react": "^19.2.0",
     "react-dom": "^19.2.0",
     "zod": "^3.25.76"
