+++ packages/ui/package.json
@@ -63,7 +63,7 @@
     "lodash": "^4.17.21",
     "lucide-react": "^0.542.0",
     "moment": "^2.30.1",
-    "next": "15.3.3",
+    "next": "16.0.1",
     "next-themes": "^0.4.6",
     "react": "^19.2.0",
     "react-colorful": "^5.6.1",
